#1804 Provided separate content builder method for description column to escape all html tags.
